requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# yacu-discord-monitor/main.py: 1
discord.py == 1.3.4

# yacu-discord-monitor/main.py: 3
google_api_python_client == 1.9.3

# yacu-discord-monitor/main.py: 2
googleapis_common_protos == 1.52.0
